+++ retz-server/src/test/java/io/github/retz/web/WebConsoleTest.java
@@ -82,6 +82,7 @@ public class WebConsoleTest {
         webConsole = new WebConsole(config);
         WebConsole.setScheduler(scheduler);
         awaitInitialization();
+
         webClient = new Client(config.getUri(), config.checkCert());
     }